--- src/main/java/Parent.java
@@ -1,8 +1,0 @@
-import org.testng.annotations.BeforeSuite;
-
-public class Parent {
-    @BeforeSuite(alwaysRun = true)
-    public void beforeSuite(){
-        System.out.println("Before suite");
-    }
-}
