+++ src/test/java/org/academy/lesson1/OpenMainUrlTest.java
@@ -1,0 +1,23 @@
+package org.academy.lesson1;
+
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.testng.annotations.Test;
+
+public class OpenMainUrlTest {
+    @Test
+    public void openUrlTest() {
+        System.setProperty("webdriver.chrome.driver", "src/main/resources/chromedriver");
+        WebDriver driver = new ChromeDriver();
+
+        driver.manage().window().maximize();
+
+        driver.get("https://github.com/");
+        try {
+            Thread.sleep(2000);
+        } catch (InterruptedException e) {
+            e.printStackTrace();
+        }
+        driver.close();
+    }
+}
