+++ src/test/java/org/academy/lesson1/SimpleTestNG3.java
@@ -1,0 +1,13 @@
+package org.academy.lesson1;
+
+import org.testng.annotations.Listeners;
+import org.academy.TestNgListener;
+import org.testng.annotations.Test;
+
+@Listeners(TestNgListener.class)
+public class SimpleTestNG3 {
+    @Test
+    public void testListener() {
+        System.out.println("Verify Listener TestNg");
+    }
+}
