+++ testngxml/beforesuite.xml
@@ -1,14 +1,9 @@
 <!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd">
 <suite name="Suite" parallel="false">
     <test name="Test1">
-        <groups>
-<!--            <run>-->
-<!--                <include name="academy"/>-->
-<!--            </run>-->
-        </groups>
         <classes>
-            <class name="lesson1.SimpleTestNG1" />
-            <class name="lesson1.SimpleTestNG2" />
+            <class name="org.academy.lesson1.SimpleTestNG1" />
+            <class name="org.academy.lesson1.SimpleTestNG2" />
         </classes>
     </test>
