+++ testngxml/testng.xml
@@ -2,7 +2,7 @@
 <suite name="Test suite">
     <test name="Tests">
         <packages>
-            <package name="lesson1"/>
+            <package name="org.academy.lesson1"/>
         </packages>
     </test>
 </suite>
